This adds the examples to the windows make file and allows them to compile.
Some issues:
1. the ability to use the conf file
Not sure why yet but the conf file is not been taken into account on windows

2. the ability to use the env variables to set a default appender defined in the examples library.
I think this is because as I call log4c_init() first, the appender, s13_stderr say, gets created with default
values and becomes the default appender for the root category.  When I call the examples init routine it creates
the s13_stderr appender with the right settings, and this ousts the existing s13_stderr appender in the
appender types hash....but I guess the root category is still pointing to the original s13_stderr appender.
Calling the examaples lib init first doesn't work either because I think when you call log4c_init() it creates the appender factory which creates a new appender hash
so the appender types created by the examples lib are lost.  Or something like that.....